Move files from testing/integration_testing to scripts/gha
This moves the integration testing scripts to the location where all the other scripts are being stored, for consistency and reuse.

The two entry points for the tools (restore_secrets.py and build_testapps.py) are moved to the top directory scripts/gha, while the rest of the files are moved to scripts/gha/integration_testing to avoid cluttering the directory.
